fix: FAPI 1.0 Advanced OIDF適合性テスト不合格項目の修正

+++ libs/idp-server-core-adapter/src/main/java/org/idp/server/core/adapters/datasource/authentication/transaction/command/MysqlExecutor.java
@@ -20,6 +20,7 @@ import java.util.ArrayList;
 import java.util.List;
 import org.idp.server.core.openid.authentication.AuthenticationTransaction;
 import org.idp.server.core.openid.authentication.AuthenticationTransactionIdentifier;
+import org.idp.server.core.openid.authentication.AuthorizationIdentifier;
 import org.idp.server.core.openid.identity.User;
 import org.idp.server.core.openid.identity.device.AuthenticationDevice;
 import org.idp.server.platform.datasource.SqlExecutor;
@@ -184,6 +185,24 @@ public class MysqlExecutor implements AuthenticationTransactionCommandSqlExecuto
         """
         DELETE FROM authentication_transaction
         WHERE id = ?
+        AND tenant_id = ?
+    """;
+
+    List<Object> params = new ArrayList<>();
+    params.add(identifier.value());
+    params.add(tenant.identifier().value());
+
+    sqlExecutor.execute(sqlTemplate, params);
+  }
+
+  @Override
+  public void deleteByAuthorizationIdentifier(Tenant tenant, AuthorizationIdentifier identifier) {
+    SqlExecutor sqlExecutor = new SqlExecutor();
+
+    String sqlTemplate =
+        """
+        DELETE FROM authentication_transaction
+        WHERE authorization_id = ?
         AND tenant_id = ?
     """;
